order-service/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
PyJWT==2.8.0
pika==1.3.2
requests==2.31.0
Werkzeug==2.3.0
pybreaker==1.0.1
tenacity==8.2.3